renamed some stuff. builds and runs
