+++ src/OpenKNX/Facade.h
@@ -11,7 +11,7 @@
 #include "OpenKNX/Time/TimeManager.h"
 #include "OpenKNX/Time/TimeProvider.h"
 #include "OpenKNX/Time/Calendar.h"
-#include "OpenKNX/IO/IO.h"
+#include "OpenKNX/GPIO/GPIO.h"
 #ifdef ParamBASE_Latitude
         #include "OpenKNX/Sun/SunCalculation.h"
 #endif
@@ -88,7 +88,7 @@ namespace OpenKNX
         Modules modules;
         Flash::Driver openknxFlash;
         Flash::Driver knxFlash;
-        IO::GPIOHAL openknxGPIO;
+        GPIO::Manager gpio;
 
         void init(uint8_t firmwareRevision);
         void loop();
